+++ backend/src/main/java/com/tamnara/backend/news/repository/NewsImageRepository.java
@@ -1,6 +1,6 @@
-package com.tamnara.backend.repository;
+package com.tamnara.backend.news.repository;
 
-import com.tamnara.backend.domain.NewsImage;
+import com.tamnara.backend.news.domain.NewsImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
