+++ backend/src/main/java/com/tamnara/backend/news/repository/NewsTagRepository.java
@@ -1,11 +1,8 @@
-package com.tamnara.backend.repository;
+package com.tamnara.backend.news.repository;
 
-import com.tamnara.backend.domain.NewsTag;
+import com.tamnara.backend.news.domain.NewsTag;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
 
 @Repository
 public interface NewsTagRepository extends JpaRepository<NewsTag, Long> {
