+++ backend/src/main/java/com/tamnara/backend/news/repository/TimelineCardRepository.java
@@ -1,10 +1,8 @@
-package com.tamnara.backend.repository;
+package com.tamnara.backend.news.repository;
 
-import com.tamnara.backend.domain.TimelineCard;
+import com.tamnara.backend.news.domain.TimelineCard;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
 
 @Repository
 public interface TimelineCardRepository extends JpaRepository<TimelineCard, Long> {
